+++ src/DocumentManager.h
@@ -14,8 +14,8 @@ public:
     void                    AddDocumentAtEnd(CDocument doc);
     void                    RemoveDocument(int index);
     void                    SetDocument(int index, CDocument doc) { m_documents[index] = doc; }
-    size_t                  GetCount() const { return m_documents.size(); }
-    size_t                  GetIndexForPath(const std::wstring& path) const;
+    int                     GetCount() const { return (int)m_documents.size(); }
+    int                     GetIndexForPath(const std::wstring& path) const;
     CDocument               GetDocument(int index) const { return m_documents.at(index); }
     Document                GetScintillaDocument(int index) const { return m_documents.at(index).m_document; }
     void                    ExchangeDocs(int src, int dst);
